Release 1.2.0 - Added Terms and License - Addded about - Added help - Updated welcome screen - Users can now add custom levels - Added option to disable some security features - Countdown is now updated on millisecond precision - Bug Fixes

+++ wikiLYNX/include/mainwindow.h
@@ -31,7 +31,7 @@ class MainWindow : public QMainWindow
     std::string domain;
     int endTime = 0;
     int cCount = 0;
-    int countdown = 0;
+    float countdown = 0;
     std::string aTime;
     std::string cTime;
     QTimer *timer = new QTimer(this);
